Reorganize ee directory (#7964)
* Re-standardize ee directory layout

* Fix spelling
