+++ apps/prairielearn/src/ee/auth/saml/router.ts
@@ -8,7 +8,7 @@ import * as authnLib from '../../../lib/authn';
 
 import { strategy } from './index';
 import { SamlTest } from './router.html';
-import { getInstitutionSamlProvider } from '../../institution/utils';
+import { getInstitutionSamlProvider } from '../../lib/institution';
 
 const router = Router({ mergeParams: true });
 
@@ -20,7 +20,7 @@ router.get('/login', (req, res, next) => {
     additionalParams: req.query.RelayState
       ? {
           // This is used be the SAML configuration page to test SAML. It includes
-          // `?RelayState=test` in the login request. When the callback page recieves
+          // `?RelayState=test` in the login request. When the callback page receives
           // that value, it displays the received attributes instead of creating a
           // new session for the user.
           RelayState: req.query.RelayState,
