+++ apps/prairielearn/src/ee/lib/institution.sql
