+++ apps/prairielearn/src/ee/pages/institutionAdminCourse/institutionAdminCourse.html.ts
