+++ apps/prairielearn/src/ee/pages/institutionAdminCourse/institutionAdminCourse.sql
