+++ apps/prairielearn/src/ee/pages/institutionAdminCourseInstance/institutionAdminCourseInstance.html.ts
