+++ apps/prairielearn/src/ee/pages/institutionAdminCourseInstance/institutionAdminCourseInstance.sql
