+++ apps/prairielearn/src/ee/pages/institutionAdminCourses/institutionAdminCourses.html.ts
