+++ apps/prairielearn/src/ee/pages/institutionAdminCourses/institutionAdminCourses.sql
